rename: myReviewCard => reviewCard
